+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial
 
-## [1.0.0-ea] - 2025-6-05
+## [1.0.0-ea] - 2025-06-05
 - Update CODEOWNERS
 - Updated `getAuthorization` to use the correct API URL.
 - Rename `getConnection(name: string)` -> `getAuthorization(developerName: string, attachmentNameOrColorUrl = "HEROKU_APPLINK")`, accepting a new attachmentNameOrColorOrUrl to use a specific Applink addon's config.
 - Remove node-fetch in favor of native fetch, add `HTTPResponseError`
 
-## Unreleased
+## [1.0.0-ea.0] - 2025-06-06
 - Add `X-App-UUID` header, `heroku-applink-node-sdk` UA.